Broken-stick (segmented) approaches to estimating SM50

A wrapper function allowing for multiple methods of broken-stick regression to be applied using a standard format for inputs. See vignette("broken-stick") for more information.

Usage

broken_stick(
  dat,
  xvar,
  yvar,
  verbose = FALSE,
  ci = 95,
  lower = NULL,
  upper = NULL,
  trans = "none",
  method = c("segmented", "chngpt", "regrans", "stevens", "all")
)

Arguments

dat

data frame or matrix containing the data

xvar

Name of column (integer or double) of measurements for the x-axis variable (e.g., carapace width).

yvar

Name of column (integer or double) of measurements for the y-axis variable (e.g., claw height).

verbose

Boolean; Should the standard error, confidence intervals, etc. be returned, or just the estimate of SM50?

ci

Integer; type of confidence intervals to return for SM50, defaults to 95%.

lower

Integer or double; the lower bound for possible SM50 values. Must be on the same trans of the data. Defaults to the 20th percentile of the x-variable.

upper

Integer or double; the upper bound for possible SM50 values. Must be on the same trans of the data. Defaults to the 80th percentile of the x-variable.

trans

Transformation to be applied to the data before performing the regression: "none", "log" (both variables are log-transformed), or "std" (both variables are standardized = scaled and centered). If no string is provided, no transformation is performed (i.e., the default is "none").

method

Method to use for the regression. A single string or string vector containing one or more of c("segmented", "chngpt", "regrans", "stevens"), or "all" to return the results of all methods for comparison.

Value

If verbose is FALSE (the default), an estimate of SM50 from the specified method(s). Otherwise, output is a list that also includes the standard error and confidence intervals for the SM50 estimate as well as the model object(s) and/or regression parameters for the lines in the broken-stick models.

Examples

set.seed(12)
fc <- fake_crustaceans(n=100, L50=100, allo_params=c(1, 0.2, 1.1, 0.2))
broken_stick(fc, xvar="x", yvar="y", method=c("segmented", "chngpt"))
#>    chngpt segmented 
#>  89.44561  88.71463