Classification approaches to estimating SM50

Classification approaches to estimating SM50

Usage

cluster_mods(
  dat,
  xvar,
  yvar,
  log = FALSE,
  method = c("mclust", "Somerton", "kmeans", "hclust", "infl_pt", "all")
)

Arguments

dat

data frame or matrix containing the data

xvar

Name of column (integer or double) of measurements for the x-axis variable (e.g., carapace width).

yvar

Name of column (integer or double) of measurements for the y-axis variable (e.g., claw height).

log

Boolean; should both variables be log-transformed before performing the regression? Defaults to FALSE.

method

Classification method to use. A single string or vector containing one or more of c("mclust", "Somerton", "kmeans", "hclust", "infl_pt"), or "all" to return the results of all methods for comparison.

Value

An estimate of SM50 from the specified method(s).

Examples

set.seed(12)
fc <- fake_crustaceans(n = 100, L50 = 100, allo_params = c(1, 0.2, 1.1, 0.2))
cluster_mods(fc, xvar = "x", yvar = "y", method = c("kmeans"))
#>          xvar     yvar
#> 1    75.38865 15.23920
#> 2   136.54339 45.87786
#> 3    85.86511 16.92335
#> 4    86.59990 16.94732
#> 5    65.04716 13.22994
#> 6    99.55408 32.83798
#> 7    98.69303 19.95659
#> 8    92.43490 16.95435
#> 9   102.87072 33.10017
#> 10  113.56030 37.01815
#> 11   89.44561 16.65966
#> 12   79.12235 16.67990
#> 13   89.40867 16.59422
#> 14  105.23904 33.84619
#> 15  101.95168 19.91354
#> 16   90.93071 17.80021
#> 17  128.77758 40.23859
#> 18  111.81025 38.40130
#> 19  115.13936 36.00876
#> 20   99.13390 20.54175
#> 21  109.47283 34.12359
#> 22  145.14403 47.01046
#> 23  125.23958 39.72831
#> 24   98.95082 30.88344
#> 25   84.49510 17.39401
#> 26   99.65230 20.61357
#> 27  101.01789 19.53285
#> 28  107.62245 34.70028
#> 29  107.91600 37.02303
#> 30  112.24129 34.84253
#> 31  118.47962 35.78884
#> 32  146.44072 45.60869
#> 33   94.17943 18.30612
#> 34   83.59016 16.46103
#> 35   97.55087 19.09151
#> 36   95.29717 18.30308
#> 37  110.49568 35.19487
#> 38   95.40975 19.62598
#> 39  120.96211 41.17689
#> 40   84.91098 17.85249
#> 41  107.09968 36.41382
#> 42   81.88014 16.17739
#> 43  116.56269 40.69230
#> 44   73.08749 14.94097
#> 45   98.82993 31.92601
#> 46  113.98932 37.07901
#> 47   85.45893 17.69469
#> 48  108.79996 22.12360
#> 49  119.62907 36.32811
#> 50   95.14802 18.34427
#> 51  104.14630 18.36215
#> 52  102.74659 32.58723
#> 53  114.13654 38.72763
#> 54  145.40670 45.19577
#> 55   83.98220 25.13179
#> 56  119.69304 41.47791
#> 57  115.78499 37.62930
#> 58   78.71454 22.97434
#> 59   99.99923 20.76300
#> 60  111.28409 35.69627
#> 61  113.13093 36.73550
#> 62  124.88841 40.34772
#> 63  122.11537 39.69475
#> 64  108.94258 33.34561
#> 65  121.68650 39.48582
#> 66  121.93580 41.68790
#> 67  144.08211 46.75849
#> 68   62.01480 13.20176
#> 69  124.42241 39.51942
#> 70  127.90123 40.06073
#> 71   94.49199 31.06592
#> 72  110.00640 33.61806
#> 73   96.41187 19.98987
#> 74  101.34961 19.74040
#> 75  102.93379 20.95561
#> 76   92.32324 18.15865
#> 77   79.57892 15.65607
#> 78   97.32099 19.78192
#> 79  115.33512 39.78074
#> 80  101.44063 30.70302
#> 81  105.08516 20.90106
#> 82   79.51881 15.62635
#> 83  100.95779 32.92634
#> 84  128.28932 38.88932
#> 85  104.53241 20.31787
#> 86  122.94313 37.65059
#> 87  101.46551 19.94880
#> 88  127.27418 43.76838
#> 89   94.16222 18.31404
#> 90   85.73203 16.49705
#> 91  112.52897 36.28876
#> 92   85.30652 16.89971
#> 93  122.95119 40.92803
#> 94  107.58525 35.14126
#> 95  125.67406 40.66184
#> 96   98.15421 20.05136
#> 97  114.04563 35.65158
#> 98   91.10524 19.66188
#> 99  100.21973 33.14270
#> 100  84.85402 17.61139