morphmat: Modeling Size at Maturity Using Morphometric Data

A compilation of methods used to estimate size at (sexual) maturity based on morphometric data, most commonly applied to crabs, lobsters, and other crustaceans. Approaches include modeling approaches based on piecewise/segmentedlinear regression as well as numerous clustering-based methods.

See also

Author

Maintainer: Ruby Krasnow ruby.krasnow@maine.edu (ORCID)