REGRANS broken-stick regression method

REGRANS broken-stick regression method

Usage

regrans_fun(
  x,
  y,
  min_changept = min(x, na.rm = TRUE),
  max_changept = max(x, na.rm = TRUE),
  n_tries = 100
)

Arguments

x

Integer or double vector of measurements for the x-axis variable (e.g., carapace width).

y

Integer or double vector of measurements for the y-axis variable (e.g., claw height).

min_changept

Integer or double; the lower bound for possible SM50 values. Must be on the same scale of the data. Defaults to the minimum value of the x-variable.

max_changept

Integer or double; the upper bound for possible SM50 values. Must be on the same scale of the data. Defaults to the maximum value of the x-variable.

n_tries

Number of breakpoints to test within the unknown range.

Value

Data frame with breakpoints tested and their sum of squares

Examples

fc <- fake_crabs(n=100, allo_params=c(1, 0.2, 1.1, 0.2))
regrans_fun(fc$x, fc$y)
#>      changept   sum_sq
#> 1    55.14267 1740.302
#> 2    56.18517 1714.415
#> 3    57.22767 1701.996
#> 4    58.27016 1693.894
#> 5    59.31266 1676.504
#> 6    60.35516 1656.482
#> 7    61.39765 1643.079
#> 8    62.44015 1626.842
#> 9    63.48265 1608.830
#> 10   64.52514 1595.645
#> 11   65.56764 1582.010
#> 12   66.61014 1565.981
#> 13   67.65264 1553.532
#> 14   68.69513 1544.766
#> 15   69.73763 1538.582
#> 16   70.78013 1534.196
#> 17   71.82262 1531.070
#> 18   72.86512 1528.833
#> 19   73.90762 1527.232
#> 20   74.95011 1524.100
#> 21   75.99261 1521.640
#> 22   77.03511 1519.874
#> 23   78.07760 1517.714
#> 24   79.12010 1515.156
#> 25   80.16260 1513.254
#> 26   81.20509 1512.581
#> 27   82.24759 1513.851
#> 28   83.29009 1516.721
#> 29   84.33259 1520.916
#> 30   85.37508 1525.859
#> 31   86.41758 1532.073
#> 32   87.46008 1539.298
#> 33   88.50257 1546.820
#> 34   89.54507 1555.268
#> 35   90.58757 1564.659
#> 36   91.63006 1574.635
#> 37   92.67256 1585.710
#> 38   93.71506 1596.359
#> 39   94.75755 1608.942
#> 40   95.80005 1624.755
#> 41   96.84255 1641.084
#> 42   97.88504 1657.266
#> 43   98.92754 1674.348
#> 44   99.97004 1688.324
#> 45  101.01254 1700.864
#> 46  102.05503 1713.589
#> 47  103.09753 1723.844
#> 48  104.14003 1732.155
#> 49  105.18252 1737.612
#> 50  106.22502 1740.045
#> 51  107.26752 1740.199
#> 52  108.31001 1739.272
#> 53  109.35251 1737.805
#> 54  110.39501 1735.935
#> 55  111.43750 1733.743
#> 56  112.48000 1731.841
#> 57  113.52250 1729.894
#> 58  114.56499 1728.348
#> 59  115.60749 1727.004
#> 60  116.64999 1725.592
#> 61  117.69248 1724.280
#> 62  118.73498 1723.247
#> 63  119.77748 1722.260
#> 64  120.81998 1721.702
#> 65  121.86247 1721.424
#> 66  122.90497 1721.545
#> 67  123.94747 1721.846
#> 68  124.98996 1722.248
#> 69  126.03246 1722.767
#> 70  127.07496 1723.008
#> 71  128.11745 1722.826
#> 72  129.15995 1722.675
#> 73  130.20245 1722.514
#> 74  131.24494 1722.373
#> 75  132.28744 1722.261
#> 76  133.32994 1722.192
#> 77  134.37243 1722.185
#> 78  135.41493 1722.284
#> 79  136.45743 1722.461
#> 80  137.49993 1722.705
#> 81  138.54242 1723.040
#> 82  139.58492 1723.494
#> 83  140.62742 1724.104
#> 84  141.66991 1724.916
#> 85  142.71241 1725.982
#> 86  143.75491 1727.355
#> 87  144.79740 1729.071
#> 88  145.83990 1728.832
#> 89  146.88240 1728.832
#> 90  147.92489 1728.832
#> 91  148.96739 1728.832
#> 92  150.00989 1728.832
#> 93  151.05238 1728.832
#> 94  152.09488 1728.832
#> 95  153.13738 1728.832
#> 96  154.17988 1728.832
#> 97  155.22237 1728.832
#> 98  156.26487 1728.832
#> 99  157.30737 1728.832
#> 100 158.34986 1740.302